@@ -1,6 +1,6 @@
 {
 	"name": "svol",
-	"version": "0.0.1",
+	"version": "0.0.2",
 	"license": "MIT",
 	"scripts": {
 		"dev": "vite dev",
@@ -26,38 +26,40 @@
 		"!dist/**/*.spec.*"
 	],
 	"peerDependencies": {
-		"svelte": "^4.0.0"
+		"svelte": "^5.16.0"
 	},
 	"devDependencies": {
-		"@sveltejs/adapter-auto": "^3.0.0",
-		"@sveltejs/kit": "^2.0.0",
-		"@sveltejs/package": "^2.0.0",
-		"@sveltejs/vite-plugin-svelte": "^3.0.0",
+		"@sveltejs/adapter-auto": "^3.3.1",
+		"@sveltejs/kit": "^2.15.0",
+		"@sveltejs/package": "^2.3.7",
+		"@sveltejs/vite-plugin-svelte": "^5.0.0",
 		"@types/proj4": "^2.5.5",
-		"autoprefixer": "^10.4.19",
+		"autoprefixer": "^10.4.20",
+		"bits-ui": "1.0.0-next.74",
+		"clsx": "^2.1.1",
 		"install": "^0.13.0",
-		"postcss": "^8.4.38",
-		"prettier": "^3.1.1",
-		"prettier-plugin-svelte": "^3.1.2",
-		"publint": "^0.1.9",
-		"svelte": "^4.2.7",
-		"svelte-check": "^3.6.0",
-		"tailwindcss": "^3.4.4",
-		"tslib": "^2.4.1",
-		"typescript": "^5.0.0",
-		"vite": "^5.0.11",
-		"vitest": "^1.2.0"
+		"lucide-svelte": "^0.469.0",
+		"paneforge": "1.0.0-next.2",
+		"postcss": "^8.4.47",
+		"prettier": "^3.3.3",
+		"prettier-plugin-svelte": "^3.2.7",
+		"publint": "^0.2.11",
+		"svelte": "^5.0.0",
+		"svelte-check": "^4.0.4",
+		"tailwind-merge": "^2.5.3",
+		"tailwind-variants": "^0.3.0",
+		"tailwindcss": "^3.4.13",
+		"tailwindcss-animate": "^1.0.7",
+		"tslib": "^2.7.0",
+		"typescript": "^5.6.2",
+		"vite": "^6.0.6",
+		"vitest": "^2.1.2"
 	},
 	"svelte": "./dist/index.js",
 	"types": "./dist/index.d.ts",
 	"type": "module",
 	"dependencies": {
-		"clsx": "^2.1.1",
-		"lucide-svelte": "^0.395.0",
-		"ol": "^9.2.4",
-		"paneforge": "^0.0.5",
-		"proj4": "^2.11.0",
-		"tailwind-merge": "^2.3.0",
-		"tailwind-variants": "^0.2.1"
+		"ol": "^10.2.1",
+		"proj4": "^2.12.1"
 	}
 }
