@@ -2,21 +2,18 @@
 	import * as ResizablePrimitive from "paneforge";
 	import { cn } from "$lib/utils.js";
 
-	type $$Props = ResizablePrimitive.PaneGroupProps;
-
-	let className: $$Props["class"] = undefined;
-	export let direction: $$Props["direction"];
-	export let paneGroup: $$Props["paneGroup"] = undefined;
-	export let el: $$Props["el"] = undefined;
-	export { className as class };
+	let {
+		ref = $bindable(null),
+		this: paneGroup = $bindable(),
+		class: className,
+		...restProps
+	}: ResizablePrimitive.PaneGroupProps & {
+		this?: ResizablePrimitive.PaneGroup;
+	} = $props();
 </script>
 
 <ResizablePrimitive.PaneGroup
-	bind:el
-	bind:paneGroup
-	{direction}
+	bind:this={paneGroup}
 	class={cn("flex h-full w-full data-[direction=vertical]:flex-col", className)}
-	{...$$restProps}
->
-	<slot />
-</ResizablePrimitive.PaneGroup>
+	{...restProps}
+/>
