@@ -1,5 +1,7 @@
 <script>
 	import '../app.css';
+	/** @type {{children?: import('svelte').Snippet}} */
+	let { children } = $props();
 </script>
 
-<slot />
+{@render children?.()}
