@@ -1,10 +1,13 @@
 <script lang="ts">
 	import { ResizableOpenlayersMap, createMapStore } from '$lib/olmap/index.js';
+	import Expand from '$lib/mapui/menu/Expand.svelte';
 	import { Map, View } from 'ol';
 	import { onMount } from 'svelte';
 	import { type Writable } from 'svelte/store';
 	import TileLayer from 'ol/layer/Tile.js';
 	import OSM from 'ol/source/OSM.js';
+	import BboxExport from '$lib/bboxexport/BboxExport.svelte';
+	import WidgetGroup from '$lib/mapui/group/WidgetGroup.svelte';
 
 	let mapStore: Writable<Map> = $state();
 	let pageMounted = $state(false);
@@ -30,20 +33,30 @@
 <div class="h-full w-full">
 	<div class="h-[100%]">
 		{#if pageMounted}
-			<ResizableOpenlayersMap {mapStore} resizable={true} headerPane={true} dataPane={true}>
+			<ResizableOpenlayersMap
+				{mapStore}
+				resizable={true}
+				mapSize={60}
+				headerSize={5}
+				dataSize={35}
+				headerPane={true}
+				dataPane={true}
+			>
 				{#snippet header()}
-								<div
-						
-						class="h-full bg-slate-700 flex items-center justify-center shadow border-collapse"
-					>
-						<span class="font-semibold text-white">Header</span>
+					<div class="h-full bg-slate-700 flex items-center justify-center shadow border-collapse">
+						<span class="font-semibold text-white">BBOX export</span>
 					</div>
-							{/snippet}
+				{/snippet}
+				{#snippet map()}
+					<WidgetGroup position="top-right">
+						<BboxExport projections={[{ value: 'EPSG:28992', label: 'EPSG:28992' }]}></BboxExport>
+					</WidgetGroup>
+				{/snippet}
 				{#snippet data()}
-								<div  class="flex h-full items-center justify-center p-6">
+					<div class="flex h-full items-center justify-center p-6">
 						<span class="font-semibold">Show Feature Data Here</span>
 					</div>
-							{/snippet}
+				{/snippet}
 			</ResizableOpenlayersMap>
 		{/if}
 	</div>
